refactor(commit-msg): migrate remaining scaffold logic to the hook scaffolder

+++ src/commit-msg/scaffolder.js
@@ -1,5 +1,10 @@
-import {scaffold as createHook} from '../hook/index.js';
+import {exists as configExistsFor} from '@form8ion/config-file';
 
-export default function ({projectRoot}) {
-  return createHook({projectRoot, hookName: 'commit-msg', script: 'npx --no-install commitlint --edit $1'});
+import {scaffold as createHook} from '../hook/index.js';
+import commitMsgHookAlreadyDefined from './tester.js';
+
+export default async function ({projectRoot}) {
+  if (await configExistsFor({name: 'commitlint'}) && !(await commitMsgHookAlreadyDefined({projectRoot}))) {
+    await createHook({projectRoot, hookName: 'commit-msg', script: 'npx --no-install commitlint --edit $1'});
+  }
 }
